+++ projects/ngx-webrtc-lib/src/lib/components/remote-stream-view/remote-stream-view.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { RemoteStreamViewComponent } from './remote-stream-view.component';
+
+describe('RemoteStreamViewComponent', () => {
+  let component: RemoteStreamViewComponent;
+  let fixture: ComponentFixture<RemoteStreamViewComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [RemoteStreamViewComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(RemoteStreamViewComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});